requirements.txt
fastapi~=0.115.3
starlette~=0.41.0
aiofiles~=24.1.0
python-dotenv~=1.0.1
pytubefix~=8.1.1
uvicorn